cli/pyproject.toml
[project]
name = "varitykit"
version = "1.2.14"
description = "Deploy any app in 60 seconds. 60-80% cheaper than AWS."
readme = "README.md"
authors = [
    {name = "Varity Team", email = "hello@varity.com"}
]
license = {text = "MIT"}
requires-python = ">=3.8"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Build Tools",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
keywords = ["varity", "deploy", "hosting", "cli", "saas"]
dependencies = [
    # Core CLI dependencies
    "click>=8.1.0",
    "docker>=6.0.0",
    "pyyaml>=6.0",
    "requests>=2.31.0",
    "rich>=13.0.0",
    "toml>=0.10.2",
    "jinja2>=3.1.0",
    "python-dotenv>=1.0.0",

    # Security: Reasonable version constraints
    "aiohttp>=3.9.0",
    "authlib>=1.3.0",
    "ecdsa>=0.18.0",
    "python-jose>=3.3.0",
    "PyJWT>=2.8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "ruff>=0.1.0",
    "mypy>=1.7.0",
    "types-requests>=2.31.0",
    "types-toml>=0.10.0",
    "types-PyYAML>=6.0.0",
    "types-click>=7.1.0",
    "isort>=5.12.0",
    "autoflake>=2.2.0",
]
blockchain = [
    "web3>=6.0.0",
    "eth-account>=0.10.0",
    "eth-utils>=2.3.0",
]
all = [
    "web3>=6.0.0",
    "eth-account>=0.10.0",
    "eth-utils>=2.3.0",
    "psutil>=5.9.0",
]

[project.scripts]
varitykit = "varitykit.cli.main:cli"

[project.urls]
Homepage = "https://www.varity.so"
Documentation = "https://docs.varity.so"
Repository = "https://github.com/varity-labs/varity-sdk"
"Bug Tracker" = "https://github.com/varity-labs/varity-sdk/issues"

[build-system]
requires = ["setuptools>=68.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
license-files = []

[tool.setuptools.packages.find]
where = ["."]
include = ["varitykit*"]
exclude = ["varitykit.tests*", "varitykit_old*"]

[tool.setuptools.package-data]
"varitykit" = ["templates/**/*", "config/*.yaml"]
"varitykit.scripts" = ["*.js", "package.json"]

[tool.black]
line-length = 100
target-version = ['py38', 'py39', 'py310', 'py311', 'py312']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

[tool.ruff]
line-length = 100
target-version = "py38"
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "C",  # flake8-comprehensions
    "B",  # flake8-bugbear
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
follow_imports = "normal"
ignore_missing_imports = true

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --strict-markers --cov=varitykit --cov-report=term-missing"
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]

[tool.coverage.run]
source = ["varitykit"]
omit = ["*/tests/*", "*/__pycache__/*", "*/site-packages/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]
